Update Defining-Components.adoc
Updated the order of the cmdIn and cmdRegOut special ports to always have cmdIn before cmdRegOut. This order is given a number of times within the FPP User's Guide, as well as when code is copy-pasted for the Math Component Tutorial. Although the opposite order (cmdRegOut before cmdIn) is seemingly given when code is auto-generated for components, I would posit that it is less intuitive to place them in this latter, non-chronological order.