Revert "Revise test script"
This reverts commit 4c76a5212e7117fbf5dc41d9bee1fe68f2f9cd3a.

Reverting the change to the test script. This change was trying
to be more precise about when to replace the author text in
the comment. However, the interaction between CI and the
pattern matching does not work as expected. Reverting to the
less precise pattern. This pattern should work, because the
script is called only when the author needs to be replaced.